Split microservice
